MAINT: consolidate lib/tests/test_shape_base.py and torch_np/tests/test_shape_base.py
